app/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
redis
